requirements.txt
langchain-astradb
langchain
langchain_community
langchain-google-genai
datasets
pypdf
python-dotenv
flask
pandas
-e .